# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Brasileirão API",
        "description": "API REST para o sistema do Campeonato Brasileiro Série A",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/api/admin/games": {
            "get": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Listar jogos com paginação e filtros",
                "operationId": "ad55ab816bcc78797ac1f3c1c788cb37",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "team",
                        "in": "query",
                        "description": "Filtrar por nome do time",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Data inicial (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Data final (Y-m-d)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de jogos"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Criar jogo",
                "operationId": "d48c42d59c4dbd68a4cbe9b7ca96856e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "home_team_id",
                                    "away_team_id",
                                    "game_date"
                                ],
                                "properties": {
                                    "home_team_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "away_team_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "game_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-05-10 16:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jogo criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/games/{id}/score": {
            "patch": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Lançar placar do jogo",
                "operationId": "0876ddd7cf2ccfd441e0506e19917ff5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "home_score",
                                    "away_score"
                                ],
                                "properties": {
                                    "home_score": {
                                        "type": "integer",
                                        "example": 2,
                                        "minimum": 0
                                    },
                                    "away_score": {
                                        "type": "integer",
                                        "example": 1,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Placar registrado, status muda para realizado"
                    },
                    "404": {
                        "description": "Jogo não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "403": {
                        "description": "Acesso negado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/games/{id}": {
            "delete": {
                "tags": [
                    "Admin - Games"
                ],
                "summary": "Excluir jogo (apenas realizados nos últimos 3 dias)",
                "operationId": "f250a7e076889da52490e481d15931b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jogo removido"
                    },
                    "403": {
                        "description": "Exclusão não permitida ou acesso negado"
                    },
                    "404": {
                        "description": "Jogo não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}